+++ support/apiv2/case_attachment_client.go
@@ -44,6 +44,7 @@ var newCaseAttachmentClientHook clientHook
 // CaseAttachmentCallOptions contains the retry settings for each method of CaseAttachmentClient.
 type CaseAttachmentCallOptions struct {
 	ListAttachments []gax.CallOption
+	GetAttachment   []gax.CallOption
 }
 
 func defaultCaseAttachmentGRPCClientOptions() []option.ClientOption {
@@ -75,6 +76,7 @@ func defaultCaseAttachmentCallOptions() *CaseAttachmentCallOptions {
 				})
 			}),
 		},
+		GetAttachment: []gax.CallOption{},
 	}
 }
 
@@ -91,6 +93,7 @@ func defaultCaseAttachmentRESTCallOptions() *CaseAttachmentCallOptions {
 					http.StatusServiceUnavailable)
 			}),
 		},
+		GetAttachment: []gax.CallOption{},
 	}
 }
 
@@ -100,6 +103,7 @@ type internalCaseAttachmentClient interface {
 	setGoogleClientInfo(...string)
 	Connection() *grpc.ClientConn
 	ListAttachments(context.Context, *supportpb.ListAttachmentsRequest, ...gax.CallOption) *AttachmentIterator
+	GetAttachment(context.Context, *supportpb.GetAttachmentRequest, ...gax.CallOption) (*supportpb.Attachment, error)
 }
 
 // CaseAttachmentClient is a client for interacting with Google Cloud Support API.
@@ -140,6 +144,17 @@ func (c *CaseAttachmentClient) Connection() *grpc.ClientConn {
 // ListAttachments list all the attachments associated with a support case.
 func (c *CaseAttachmentClient) ListAttachments(ctx context.Context, req *supportpb.ListAttachmentsRequest, opts ...gax.CallOption) *AttachmentIterator {
 	return c.internalClient.ListAttachments(ctx, req, opts...)
+}
+
+// GetAttachment retrieve an attachment associated with a support case.
+//
+// EXAMPLES:
+//
+// cURL:
+//
+// Python:
+func (c *CaseAttachmentClient) GetAttachment(ctx context.Context, req *supportpb.GetAttachmentRequest, opts ...gax.CallOption) (*supportpb.Attachment, error) {
+	return c.internalClient.GetAttachment(ctx, req, opts...)
 }
 
 // caseAttachmentGRPCClient is a client for interacting with Google Cloud Support API over gRPC transport.
@@ -211,6 +226,7 @@ func NewCaseAttachmentClient(ctx context.Context, opts ...option.ClientOption) (
 		)
 
 		client.CallOptions.ListAttachments = append(client.CallOptions.ListAttachments, gax.WithClientMetrics(metrics))
+		client.CallOptions.GetAttachment = append(client.CallOptions.GetAttachment, gax.WithClientMetrics(metrics))
 	}
 
 	client.internalClient = c
@@ -302,6 +318,7 @@ func NewCaseAttachmentRESTClient(ctx context.Context, opts ...option.ClientOptio
 		)
 
 		callOpts.ListAttachments = append(callOpts.ListAttachments, gax.WithClientMetrics(metrics))
+		callOpts.GetAttachment = append(callOpts.GetAttachment, gax.WithClientMetrics(metrics))
 	}
 
 	return &CaseAttachmentClient{internalClient: c, CallOptions: callOpts}, nil
@@ -396,6 +413,30 @@ func (c *caseAttachmentGRPCClient) ListAttachments(ctx context.Context, req *sup
 	return it
 }
 
+func (c *caseAttachmentGRPCClient) GetAttachment(ctx context.Context, req *supportpb.GetAttachmentRequest, opts ...gax.CallOption) (*supportpb.Attachment, error) {
+	hds := []string{"x-goog-request-params", fmt.Sprintf("%s=%v", "name", url.QueryEscape(req.GetName()))}
+
+	hds = append(c.xGoogHeaders, hds...)
+	ctx = gax.InsertMetadataIntoOutgoingContext(ctx, hds...)
+	if gax.IsFeatureEnabled("TRACING") || gax.IsFeatureEnabled("LOGGING") {
+		ctx = callctx.WithTelemetryContext(ctx, "resource_name", fmt.Sprintf("//cloudsupport.googleapis.com/%v", req.GetName()))
+	}
+	if gax.IsFeatureEnabled("METRICS") || gax.IsFeatureEnabled("TRACING") || gax.IsFeatureEnabled("LOGGING") {
+		ctx = callctx.WithTelemetryContext(ctx, "rpc_method", "google.cloud.support.v2.CaseAttachmentService/GetAttachment")
+	}
+	opts = append((*c.CallOptions).GetAttachment[0:len((*c.CallOptions).GetAttachment):len((*c.CallOptions).GetAttachment)], opts...)
+	var resp *supportpb.Attachment
+	err := gax.Invoke(ctx, func(ctx context.Context, settings gax.CallSettings) error {
+		var err error
+		resp, err = executeRPC(ctx, c.caseAttachmentClient.GetAttachment, req, settings.GRPC, c.logger, "GetAttachment")
+		return err
+	}, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return resp, nil
+}
+
 // ListAttachments list all the attachments associated with a support case.
 func (c *caseAttachmentRESTClient) ListAttachments(ctx context.Context, req *supportpb.ListAttachmentsRequest, opts ...gax.CallOption) *AttachmentIterator {
 	it := &AttachmentIterator{}
@@ -472,4 +513,67 @@ func (c *caseAttachmentRESTClient) ListAttachments(ctx context.Context, req *sup
 	it.pageInfo.Token = req.GetPageToken()
 
 	return it
+}
+
+// GetAttachment retrieve an attachment associated with a support case.
+//
+// EXAMPLES:
+//
+// cURL:
+//
+// Python:
+func (c *caseAttachmentRESTClient) GetAttachment(ctx context.Context, req *supportpb.GetAttachmentRequest, opts ...gax.CallOption) (*supportpb.Attachment, error) {
+	baseUrl, err := url.Parse(c.endpoint)
+	if err != nil {
+		return nil, err
+	}
+	baseUrl.Path += fmt.Sprintf("/v2/%v", req.GetName())
+
+	params := url.Values{}
+	params.Add("$alt", "json;enum-encoding=int")
+
+	baseUrl.RawQuery = params.Encode()
+
+	// Build HTTP headers from client and context metadata.
+	hds := []string{"x-goog-request-params", fmt.Sprintf("%s=%v", "name", url.QueryEscape(req.GetName()))}
+
+	hds = append(c.xGoogHeaders, hds...)
+	hds = append(hds, "Content-Type", "application/json")
+	headers := gax.BuildHeaders(ctx, hds...)
+	if gax.IsFeatureEnabled("TRACING") || gax.IsFeatureEnabled("LOGGING") {
+		ctx = callctx.WithTelemetryContext(ctx, "resource_name", fmt.Sprintf("//cloudsupport.googleapis.com/%v", req.GetName()))
+	}
+	if gax.IsFeatureEnabled("METRICS") || gax.IsFeatureEnabled("TRACING") || gax.IsFeatureEnabled("LOGGING") {
+		ctx = callctx.WithTelemetryContext(ctx, "rpc_method", "google.cloud.support.v2.CaseAttachmentService/GetAttachment")
+		ctx = callctx.WithTelemetryContext(ctx, "url_template", "/v2/{name=*/*/cases/*/attachments/*}")
+	}
+	opts = append((*c.CallOptions).GetAttachment[0:len((*c.CallOptions).GetAttachment):len((*c.CallOptions).GetAttachment)], opts...)
+	unm := protojson.UnmarshalOptions{AllowPartial: true, DiscardUnknown: true}
+	resp := &supportpb.Attachment{}
+	e := gax.Invoke(ctx, func(ctx context.Context, settings gax.CallSettings) error {
+		if settings.Path != "" {
+			baseUrl.Path = settings.Path
+		}
+		httpReq, err := http.NewRequest("GET", baseUrl.String(), nil)
+		if err != nil {
+			return err
+		}
+		httpReq = httpReq.WithContext(ctx)
+		httpReq.Header = headers
+
+		buf, err := executeHTTPRequest(ctx, c.httpClient, httpReq, c.logger, nil, "GetAttachment")
+		if err != nil {
+			return err
+		}
+
+		if err := unm.Unmarshal(buf, resp); err != nil {
+			return err
+		}
+
+		return nil
+	}, opts...)
+	if e != nil {
+		return nil, e
+	}
+	return resp, nil
 }